Rename file to config-schema
