+++ lib/utils/config-schema.js
