+++ lib/utils/main-loop.js
@@ -3,13 +3,13 @@ const shouldExcludeFile = require('./should-exclude-file')
 const getDiff = require('./get-diff')
 const getDetails = require('./get-details')
 const checkForBody = require('./check-for-body')
-const defaultConfig = require('./default-config')
+const configSchema = require('./config-schema')
 const generateLabel = require('./generate-label')
 
 module.exports = async (context, handler) => {
   // Grab the config file
   const configFile = await context.config('config.yml')
-  const config = await defaultConfig.validate(configFile && configFile.todo ? configFile.todo : {})
+  const config = await configSchema.validate(configFile && configFile.todo ? configFile.todo : {})
 
   const keywords = Array.isArray(config.keyword) ? config.keyword : [config.keyword]
